requirements.txt
absl-py==2.3.1
asgiref==3.9.0
astunparse==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
opencv-python==4.10.0.84
protobuf==5.29.5
Pygments==2.19.2
PyJWT==2.9.0
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0; platform_system != "Windows"
termcolor==3.1.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
gunicorn==21.2.0
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary
django-storages
boto3
python-dotenv